udpate nugets. bug fixes

+++ Jakar.Extensions/Http/StatusExtensions.cs
@@ -7,8 +7,8 @@ namespace Jakar.Extensions;
 
 public static class StatusExtensions
 {
-    public static HttpStatusCode ToStatus( this Status         code ) => (HttpStatusCode)(int)code;
-    public static Status ToStatus( this         HttpStatusCode code ) => (Status)(int)code;
+    public static HttpStatusCode ToStatus( this Status         code ) => (HttpStatusCode)code;
+    public static Status ToStatus( this         HttpStatusCode code ) => (Status)code;
 
 
     public static string ToStringFast( this HttpStatusCode status ) => status.ToStatus()
